Update main.c
fixing data type of parameters

+++ main.c
@@ -45,7 +45,7 @@ int main() {
  
 }
  
-int add(float num1, float num2){
+int add(int num1, int num2){
  
     int addtion = num1 + num2;
     printf("Addition is %d",addition);
